feat(openamt): review, rename package
